+++ spoonbill/spec.py
@@ -1,11 +1,11 @@
 import logging
 from collections import OrderedDict
-from dataclasses import asdict, dataclass, field, is_dataclass
+from dataclasses import dataclass, field, is_dataclass
 from typing import List, Mapping, Sequence
 
 from spoonbill.common import DEFAULT_FIELDS, DEFAULT_FIELDS_COMBINED
 from spoonbill.i18n import _
-from spoonbill.utils import combine_path, common_prefix, generate_table_name, get_pointer, get_root
+from spoonbill.utils import combine_path, common_prefix, generate_table_name, get_pointer
 
 LOGGER = logging.getLogger("spoonbill")
 
@@ -85,7 +85,9 @@ class Table:
                         col = Column(**col)
                     init[name] = col
                 setattr(self, attr, init)
-            cols = DEFAULT_FIELDS if self.is_root else DEFAULT_FIELDS_COMBINED
+            cols = DEFAULT_FIELDS_COMBINED
+            if self.is_root and not self.is_combined:
+                cols = DEFAULT_FIELDS
             for col in cols:
                 if col not in self.columns:
                     self.columns[col] = Column(col, "string", col)
@@ -184,14 +186,17 @@ class Table:
         :param path: The column's JSON path without array indexes
         """
         header = get_pointer(self, abs_path, path, True)
-        if header in self.columns:
-            self.columns[header].hits += 1
-        if header in self.combined_columns:
-            self.combined_columns[header].hits += 1
-        if header in self.additional_columns:
-            self.additional_columns[header].hits += 1
+        for headers in self.columns, self.combined_columns, self.additional_columns:
+            if header in headers:
+                headers[header].hits += 1
+
         if not self.is_root:
             self.parent.inc_column(abs_path, path)
+
+    def add_array(self, header):
+        self.arrays[header] = 0
+        if not self.is_root:
+            self.parent.add_array(header)
 
     def set_array(self, header, item):
         """
@@ -219,19 +224,13 @@ class Table:
         for col_name in DEFAULT_FIELDS_COMBINED:
             self.inc_column(col_name, col_name)
 
-    def dump(self):
-        data = asdict(self)
-        if data["parent"]:
-            data["parent"] = data["parent"]["name"]
-        return data
-
     def set_preview_path(self, abs_path, path, value, max_items):
         header = get_pointer(self, abs_path, path, True)
         array = self.is_array(path)
-        self.preview_rows[-1][header] = value
+        self.preview_rows_combined[-1][header] = value
         if header in self.combined_columns:
             if not array or (array and self.arrays[array] < max_items):
-                self.preview_rows_combined[-1][header] = value
+                self.preview_rows[-1][header] = value
         if not self.is_root:
             self.parent.set_preview_path(abs_path, path, value, max_items)
 
@@ -249,7 +248,5 @@ def add_child_table(table, pointer, parent_key, key):
     table_name = generate_table_name(table.name, parent_key, key)
     child_table = Table(table_name, [pointer], parent=table)
     table.child_tables.append(table_name)
-    for t in table, get_root(table):
-        t.arrays[pointer] = 0
-        t.arrays[pointer] = 0
+    table.add_array(pointer)
     return child_table