+++ tests/test_flatten.py
@@ -48,6 +48,27 @@ def test_flatten(spec_analyzed, releases):
                     count[name] += 1
 
 
+def test_flatten_row_id_parent_id_relation(spec, releases):
+    releases[0]["tender"]["items"] = releases[0]["tender"]["items"] * 6
+    releases[0]["tender"]["items"] = releases[0]["tender"]["items"] * 6
+    releases[0]["tender"]["items"][0]["additionalClassifications"] = (
+        releases[0]["tender"]["items"][0]["additionalClassifications"] * 6
+    )
+    for _ in spec.process_items(releases):
+        pass
+    options = FlattenOptions(**{"selection": {"tenders": {"split": True}}})
+    flattener = Flattener(options, spec.tables)
+    all_rows = defaultdict(list)
+    for count, flat in flattener.flatten(releases):
+        for name, rows in flat.items():
+            all_rows[name].extend(rows)
+
+    for row in all_rows["tenders_items_class"]:
+        parent_id = row["parentID"]
+        items = [i for i in all_rows["tenders_items"] if i["rowID"] == parent_id]
+        assert items
+
+
 def test_flattener_generate_count_columns(spec, releases):
     releases[0]["tender"]["items"] = releases[0]["tender"]["items"] * 6
     for _ in spec.process_items(releases):
@@ -284,3 +305,25 @@ def test_flatten_fields_compare(spec_analyzed, releases):
                             expected = JOINABLE_SEPARATOR.join(expected)
                         assert expected == value
                 counters[name] += 1
+
+
+def test_flatten_only_no_default_columns(spec_analyzed, releases):
+    options = FlattenOptions(**{"selection": {"tenders": {"split": True, "only": ["/tender/id"]}}})
+    flattener = Flattener(options, spec_analyzed.tables)
+    for _count, flat in flattener.flatten(releases):
+        for name, rows in flat.items():
+            for row in rows:
+                assert len(rows) == 1
+                assert not set(row).difference(["/tender/id"])
+
+
+def test_flatten_buyer(spec_analyzed, releases):
+    options = FlattenOptions(**{"selection": {"parties": {"split": False}}})
+    flattener = Flattener(options, spec_analyzed.tables)
+    for count, flat in flattener.flatten(releases):
+        buyer = search(f"[{count}].buyer", releases)
+        for name, rows in flat.items():
+            for row in rows:
+                if buyer:
+                    assert "/buyer/id" in row
+                    assert "/buyer/name" in row